+++ libhoomd/python/hoomd_module.cc
@@ -61,7 +61,6 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "BondedGroupData.h"
 #include "Initializers.h"
 #include "HOOMDInitializer.h"
-#include "HOOMDBinaryInitializer.h"
 #include "RandomGenerator.h"
 #include "Compute.h"
 #include "CellList.h"
@@ -93,7 +92,6 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "IMDInterface.h"
 #include "HOOMDDumpWriter.h"
 #include "POSDumpWriter.h"
-#include "HOOMDBinaryDumpWriter.h"
 #include "PDBDumpWriter.h"
 #include "MOL2DumpWriter.h"
 #include "DCDDumpWriter.h"
@@ -522,7 +520,6 @@ BOOST_PYTHON_MODULE(hoomd)
     export_RandomInitializer();
     export_SimpleCubicInitializer();
     export_HOOMDInitializer();
-    export_HOOMDBinaryInitializer();
     export_RandomGenerator();
 
     // computes
@@ -639,7 +636,6 @@ BOOST_PYTHON_MODULE(hoomd)
     export_IMDInterface();
     export_HOOMDDumpWriter();
     export_POSDumpWriter();
-    export_HOOMDBinaryDumpWriter();
     export_PDBDumpWriter();
     export_DCDDumpWriter();
     export_MOL2DumpWriter();